simple refactor of gmail to use dojo state
